+++ xknx/knxip/__init__.py
@@ -1,7 +1,7 @@
 """This package contains all methods for serialization and deserialization of KNX/IP packets."""
 # flake8: noqa
 from .body import KNXIPBody, KNXIPBodyResponse
-from .connect_request import ConnectRequest
+from .connect_request import ConnectRequest, ConnectRequestInformation
 from .connect_response import ConnectResponse
 from .connectionstate_request import ConnectionStateRequest
 from .connectionstate_response import ConnectionStateResponse
@@ -29,6 +29,7 @@ from .knxip_enum import (
     KNXIPServiceType,
     KNXMedium,
     SearchRequestParameterType,
+    TunnellingLayer,
 )
 from .routing_busy import RoutingBusy
 from .routing_indication import RoutingIndication
@@ -51,6 +52,7 @@ __all__ = [
     "KNXIPBody",
     "KNXIPBodyResponse",
     "ConnectRequest",
+    "ConnectRequestInformation",
     "ConnectResponse",
     "ConnectionStateRequest",
     "ConnectionStateResponse",
@@ -90,5 +92,6 @@ __all__ = [
     "SRP",
     "TimerNotify",
     "TunnellingAck",
+    "TunnellingLayer",
     "TunnellingRequest",
 ]